moving components around to where they're supposed to be, separated my highlights button and popup styles and specs into folder, removed unused declarations
