+++ src/app/components/Layout.tsx
@@ -1,6 +1,5 @@
 import React, { SFC } from 'react';
 import styled, { css } from 'styled-components/macro';
-import HighlightsPopUp from '../content/components/Toolbar/HighlightsPopUp';
 import ErrorBoundary from '../errors/components/ErrorBoundary';
 import ErrorModal from '../errors/components/ErrorModal';
 import theme from '../theme';
@@ -11,7 +10,6 @@ import SkipToContentWrapper from './SkipToContentWrapper';
 const Layout: SFC = ({ children }) => <SkipToContentWrapper>
   <NavBar />
   <ErrorModal />
-  <HighlightsPopUp />
   <ErrorBoundary>
     {children}
   </ErrorBoundary>
