+++ src/app/content/components/Toolbar/MyHighlights/HighlightButton.tsx
@@ -2,12 +2,12 @@ import React, { Component } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
 import styled, { css } from 'styled-components/macro';
-import highlightIcon from '../../../../assets/highlightIcon.svg';
-import * as selectors from '../../../content/highlights/selectors';
-import theme from '../../../theme';
-import { AppState, Dispatch } from '../../../types';
-import { openMyHighlights } from '../../highlights/actions';
-import { PlainButton, toolbarDefaultText, toolbarIconStyles } from './styled';
+import highlightIcon from '../../../../../assets/highlightIcon.svg';
+import theme from '../../../../theme';
+import { AppState, Dispatch } from '../../../../types';
+import { openMyHighlights } from '../../../highlights/actions';
+import * as selectors from '../../../highlights/selectors';
+import { PlainButton, toolbarDefaultText, toolbarIconStyles } from '../styled';
 
 interface Props {
   openMyHighlights: () => void;
